# backend/tests/fixtures/openapi/breaking.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TimelineEditRequest": {
        "additionalProperties": false,
        "properties": {
          "field_name": {
            "enum": [
              "title",
              "description",
              "event_type",
              "time_expression"
            ],
            "title": "Field Name",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "field_name",
          "value"
        ],
        "title": "TimelineEditRequest",
        "type": "object"
      },
      "TimelineEnvelope": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimelineVersionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "running_candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimelineVersionView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TimelineEnvelope",
        "type": "object"
      },
      "TimelineOrdering": {
        "enum": [
          "narrative",
          "story"
        ],
        "title": "TimelineOrdering",
        "type": "string"
      },
      "TimelinePreferenceRequest": {
        "additionalProperties": false,
        "properties": {
          "full_book": {
            "title": "Full Book",
            "type": "boolean"
          }
        },
        "required": [
          "full_book"
        ],
        "title": "TimelinePreferenceRequest",
        "type": "object"
      },
...
      "TimelineRunResponse": {
        "additionalProperties": false,
        "properties": {
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "novel_id": {
            "title": "Novel Id",
            "type": "integer"
          },
          "progress": {
            "additionalProperties": true,
            "title": "Progress",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "id",
          "novel_id",
          "status"
        ],
        "title": "TimelineRunResponse",
        "type": "object"
      },
...
      "TimelineVersionView": {
        "additionalProperties": false,
        "properties": {
          "aggregates": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Aggregates",
            "type": "object"
          },
          "causal_edges": {
            "items": {
              "$ref": "#/components/schemas/TimelineVisibleEdge"
            },
            "title": "Causal Edges",
            "type": "array"
          },
          "counts": {
            "$ref": "#/components/schemas/TimelineCounts"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TimelineVisibleEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "previews": {
            "items": {
              "type": "string"
            },
            "title": "Previews",
            "type": "array"
          },
          "progress": {
            "additionalProperties": true,
            "title": "Progress",
            "type": "object"
          },
          "source": {
            "$ref": "#/components/schemas/TimelineVersionSource"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "integer"
          }
        },
        "required": [
          "source",
          "version_id",
          "status"
        ],
        "title": "TimelineVersionView",
        "type": "object"
      },
      "TimelineVisibleEdge": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "edge_type": {
            "title": "Edge Type",
            "type": "string"
          },
          "source_event_id": {
            "title": "Source Event Id",
            "type": "integer"
          },
          "target_event_id": {
            "title": "Target Event Id",
            "type": "integer"
          }
        },
        "required": [
          "source_event_id",
          "target_event_id",
          "edge_type",
          "confidence"
        ],
        "title": "TimelineVisibleEdge",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI 辅助小说创作与理解平台",
    "title": "NovelMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/timeline/{novel_id}": {
      "get": {
        "operationId": "get_timeline_api_timeline__novel_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "ordering",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TimelineOrdering",
              "default": "narrative"
            }
          },
          {
            "in": "query",
            "name": "person",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Person"
            }
          },
          {
            "in": "query",
            "name": "causal",
            "required": false,
            "schema": {
              "default": false,
              "title": "Causal",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "full_book",
            "required": false,
            "schema": {
              "default": false,
              "title": "Full Book",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "chapter_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Start"
            }
          },
          {
            "in": "query",
            "name": "chapter_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter End"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineEnvelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Timeline",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/cancel": {
      "post": {
        "operationId": "cancel_api_timeline__novel_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/events/{logical_event_id}": {
      "put": {
        "operationId": "edit_event_api_timeline__novel_id__events__logical_event_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "logical_event_id",
            "required": true,
            "schema": {
              "title": "Logical Event Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelineEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit Event",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/extract": {
      "post": {
        "operationId": "extract_timeline_api_timeline__novel_id__extract_post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Extract Timeline",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/preference": {
      "put": {
        "operationId": "set_preference_api_timeline__novel_id__preference_put",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelinePreferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Preference",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/resume": {
      "post": {
        "operationId": "resume_api_timeline__novel_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resume",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/start-or-resume": {
      "post": {
        "description": "启动或恢复时间线深度分析（Phase 08）。\n\n编排（对齐 D-12 / REQ-TIME-03）:\n  1. 确保 Phase 07 active hierarchy 存在（结构底座服务时间线）\n  2. 幂等创建/恢复 durable analysis run\n  3. 后台 worker 按章抽取 → 调和 → 晋升",
        "operationId": "start_or_resume_api_timeline__novel_id__start_or_resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Or Resume",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/status": {
      "get": {
        "operationId": "status_api_timeline__novel_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Status",
        "tags": [
          "时间线"
        ]
      }
    },
    "/api/timeline/{novel_id}/versions/{version_id}": {
      "get": {
        "operationId": "get_version_api_timeline__novel_id__versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "title": "Novel Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "ordering",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TimelineOrdering",
              "default": "narrative"
            }
          },
          {
            "in": "query",
            "name": "person",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Person"
            }
          },
          {
            "in": "query",
            "name": "causal",
            "required": false,
            "schema": {
              "default": false,
              "title": "Causal",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "full_book",
            "required": false,
            "schema": {
              "default": false,
              "title": "Full Book",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "chapter_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Start"
            }
          },
          {
            "in": "query",
            "name": "chapter_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineVersionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Version",
        "tags": [
          "时间线"
        ]
      }
    }
  }
}